Refactor and simplify middleware code
Simplify the handling/calling of middlewares by composing the middleware
functions together instead of looping through an array to call them.
Managers now only need a single jobFunc, which will call all middlewares
provided when the manager was created.

The behavior for setting up a worker with custom middlewares is also changed:
If custom middlewares are provided, the default middlewares are not
included. The old behavior can be obtained with
DefaultMiddlewares().Append(customMid). Additionally, a new
NopMiddleware can be provided if removing the default middlewares is
desired.